Set appropriate builder flags in copy-like constructor
In the Builder constructor that takes an AttributionSource, it attempts
to set various fields but these are later cleared because the associated
bits are not set. This was leading to expected fields in newly created
AttributionSource objects being unset.

Previous attempts to root cause this issue attempted to determine why
appops service would drop the state of the system server uid.

Flag: EXEMPT bugfix
Fixes: 333931259
Test: Field metrics, this issue was never observed locally

Change-Id: I20e19196fff158a67beba76b40380fd08b0fe60e