[flexiglass] Dismiss bouncer when clicking on action button
This returns to the lockscreen scene when the "emergency button" is
clicked by the user.

The change helps avoid occlusion problems as described in the attached
bug and also mimics the solution done in pre-flexiglass.

Fix: 355496091
Test: unit test updated
Test: manually verified that the emergency services activity appears and
is not occluded by the bouncer - checked on first unlock after reboot
and on subsequent unlocked
Test: manually verified that the camera app, when shown after a
double-press of the power button properly brings up the bouncer and that
unlocking it returns to the camera app as it should
Flag: com.android.systemui.scene_container

Change-Id: I6284a70992fe3ab32069d1405713477fa17d1887